SB: fixing type and rollup errors
