+++ packages/component-boilerplate/package.json
@@ -7,6 +7,7 @@
   "main": "dist/index.js",
   "exports": "./dist/index.modern.js",
   "module": "dist/index.module.js",
+  "types": "dist/index.d.ts",
   "scripts": {
     "build": "microbundle",
     "test": "jest"
@@ -18,7 +19,6 @@
     "@types/debounce": "^1.2.0",
     "@types/debug": "^4.1.5",
     "@types/jest": "^26.0.20",
-    "@types/memoizee": "^0.4.5",
     "jest": "^26.6.3",
     "jest-cucumber": "^3.0.1",
     "microbundle": "^0.13.0",
@@ -29,8 +29,9 @@
     "@wry/equality": "^0.3.3",
     "debounce": "^1.2.0",
     "dom-context": "^1.0.1",
+    "fast-memoize": "^2.5.2",
+    "graphql": "^15.5.0",
     "graphql-request": "^3.4.0",
-    "history": "^5.0.0",
-    "memoizee": "^0.4.15"
+    "history": "^5.0.0"
   }
 }
