--- packages/component-boilerplate/src/hooks/graphql/GqlType.ts
@@ -1,3 +1,0 @@
-import { RequestDocument } from "graphql-request/dist/types";
-
-export type GqlType = RequestDocument;
