--- packages/component-boilerplate/src/hooks/graphql/QueryData.ts
@@ -1,11 +1,0 @@
-import { GraphQlRequestError } from "./GraphQlRequestError";
-
-export interface BaseQueryData<T = unknown> {
-  loading: boolean;
-  data?: T;
-  errors?: GraphQlRequestError<T>;
-}
-
-export type QueryData<T> = BaseQueryData<T> & {
-  refetch: () => unknown;
-};
